--- src/main/AppMain.java
@@ -1,13 +1,0 @@
-package main;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ValidatingFormInputApplication {
-
-	public static void main(String[] args) throws Exception {
-		SpringApplication.run(ValidatingFormInputApplication.class, args);
-	}
-
-}
