+++ src/main/java/com/lapenta/main/Main.java
@@ -1,0 +1,13 @@
+package com.lapenta.main;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+
+	public static void main(String[] args) throws Exception {
+		SpringApplication.run(Main.class, args);
+	}
+
+}
