+++ src/main/resources/templates/form.html
@@ -1,0 +1,25 @@
+<html>
+    <body>
+        <form action="#" th:action="@{/}" th:object="${URLForm}" method="post">
+            <table>
+                <tr>
+                    <td>URL:</td>
+                    <td><input type="text" th:field="*{longUrl}" /></td>
+                    <td th:if="${#fields.hasErrors('longUrl')}" th:errors="*{longUrl}">Name Error</td>
+                </tr>
+               
+                <tr>
+                    <td><button type="submit">Submit</button></td>
+                </tr>
+                
+                <tr>
+				<button onclick="location.href='/show_all'" type="button">Show All</button>                
+         		</tr>
+         		
+                <tr>
+				<button onclick="location.href='/most_clicked'" type="button">Most Clicked</button>                
+         		</tr>
+            </table>
+        </form>
+    </body>
+</html>
