+++ src/main/resources/templates/results.html
@@ -1,0 +1,11 @@
+<html>
+	<body>
+	<table id="users">
+		<th>Short URL :</th> 
+		<td>{{shortUrl}}</td>
+		
+		<c:set var="su" value="${shortUrl}" />
+    	<c:out value="${su}" escapeXml="false" />
+	</table>
+	</body>
+</html>
