--- src/save/Save.java
@@ -1,50 +1,0 @@
-package save;
-
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.util.HashMap;
-
-public class Save {
-
-
-	public void serialize(HashMap<String, String> hmap) {
-		try {
-			FileOutputStream fos = new FileOutputStream("hashmap.ser");
-			ObjectOutputStream oos = new ObjectOutputStream(fos);
-			oos.writeObject(hmap);
-			oos.close();
-			fos.close();
-		} catch (IOException ioe) {
-			ioe.printStackTrace();
-		}
-	}
-
-
-
-	@SuppressWarnings({ "unchecked"})
-	public HashMap<String, String> deserialize () {
-
-		HashMap<String, String> map = new HashMap<String,String>();
-	      try
-	      {
-	         FileInputStream fis = new FileInputStream("hashmap.ser");
-	         ObjectInputStream ois = new ObjectInputStream(fis);
-	         map = (HashMap<String, String>) ois.readObject();
-	         ois.close();
-	         fis.close();
-	      }catch(IOException ioe)
-	      {
-	         ioe.printStackTrace();
-	         return map;
-	      }catch(ClassNotFoundException c)
-	      {
-	         System.out.println("Class not found");
-	         c.printStackTrace();
-	         return map;
-	      }
-	      return map;
-	  }
-}
